Move qvto-transformations needed for the test on first level of testmodel.
This change of location of the two transformations (scalePopulationconditions.qvto and scalePopulation.qvto)
leads to a different structure than the action.repository.
In this situation it means, that the required transformations in the Test Action cannot be found by Maven, as:
a) Maven has different restrictions on where resources can be found than Eclipse JUnit Test execution
b) Somehow the transformations are not found if they are in the standard folder structure of the action.repository,
even though the URIs seem to be correct.

Anyways, as long as the referenced URIs are correct and on the first level of the testmodel,
the structure of the action.repository does not seem to have
a technical meaning, apart from strucuturing it and Eclipse and Maven Unit-Test Execution find the required
transformations.